# toteco_openapi.yaml
openapi: 3.0.2
info:
  title: TOTECO API
  description: API to manage publications about products (coffee, tea or toast) like a social media. The idea is to create a community that publishes photos of their toasts with coffee or tea adding the price and being able to rate the products and the establishments.
  version: "1.0"
  contact:
    email: a25540@svalero.com
servers:
  - url: http://localhost:8081
...
paths:
...
  /menus:
    get:
      tags:
        - Menu
      summary: Get all menus
      description: Get all menus
      operationId: getAllMenus
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
    post:
      security:
        - bearerAuth: []
      tags:
        - Menu
      summary: Save menu
      description: Save one menu
      operationId: saveMenu
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuDTO'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                    
    delete:
      security:
        - bearerAuth: []
      tags:
        - Menu
      summary: Delete menu
      description: Delete all menus
      operationId: deleteAllMenus
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: string
                example: All menus deleted.
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /menus/{id}:
    get:
      tags:
        - Menu
      summary: Get menu by id
      description: Get one menu by the indicated id
      operationId: getMenuById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
    
    put:
      security:
        - bearerAuth: []
      tags:
        - Menu
      summary: Update menu
      description: Update an existing menu
      operationId: updateMenu
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuDTO'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                    
    delete:
      security:
        - bearerAuth: []
      tags:
        - Menu
      summary: Delete menu by id
      description: Delete an existing menu
      operationId: deleteMenu
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: string
                example: Menu deleted.
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /menus/date/{date}:
    get:
      tags:
        - Menu
      summary: Get menu by date
      description: Get menus created at the indicated date. The date format is timestamp in miliseconds
      operationId: getMenuByDate
      parameters:
        - $ref: '#/components/parameters/date'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /menus/date/between:
    get:
      tags:
        - Menu
      summary: Get menu by date between
      description: Get menus created at the indicated dates. The dates format is timestamp in miliseconds
      operationId: getMenuByDateBetween
      parameters:
        - $ref: '#/components/parameters/minDate'
        - $ref: '#/components/parameters/maxDate'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /menus/price/{price}:
    get:
      tags:
        - Menu
      summary: Get menu by price
      description: Get menus with the price indicated
      operationId: getMenuByPrice
      parameters:
        - $ref: '#/components/parameters/price'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /menus/price/between:
    get:
      tags:
        - Menu
      summary: Get menu by price between
      description: Get menus between the two prices indicated.
      operationId: getMenuByPriceBetween
      parameters:
        - $ref: '#/components/parameters/minPrice'
        - $ref: '#/components/parameters/maxPrice'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /menus/score/{score}:
    get:
      tags:
        - Menu
      summary: Get menu by score
      description: Get menus with the score indicated. The score goes from 0 to 5.
      operationId: getMenuByScore
      parameters:
        - $ref: '#/components/parameters/score'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /menus/score/between:
    get:
      tags:
        - Menu
      summary: Get menu by score between
      description: Get menus between the two scores indicated. The score goes from 0 to 5.
      operationId: getMenuByScoreBetween
      parameters:
        - $ref: '#/components/parameters/minScore'
        - $ref: '#/components/parameters/maxScore'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
...
components:
  schemas:
...
    Menu:
      properties:
        id:
          type: integer
          format: int64
          example: 0
        date:
          type: string
          format: dd-MM-yyyy
          example: 02-04-2022
        price:
          type: number
          example: 3.4
        score:
          type: number
          example: 4.5            
    MenuDTO:
      properties:
        price:
          type: number
          example: 3.4
        score:
          type: number
          example: 4.5            
...
    Exception:
      properties:
        code: 
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
        message:
          type: string
  parameters:
    id:
      name: id
      in: path
      description: id of the element
      required: true
      schema:
        type: integer
        format: int64
        example: 0
    date:
      name: date
      in: path
      description: creation date
      required: true
      schema:
        type: integer
        example: 1640604419000
    minDate:
      name: minDate
      in: query
      description: date from in timestamp miliseconds
      required: true
      schema:
        type: integer
        example: 1640604419000
    maxDate:
      name: maxDate
      in: query
      description: date to in timestamp miliseconds
      required: true
      schema:
        type: integer
        example: 1641604419000
    price:
      name: price
      in: path
      required: true
      schema:
        type: number
        example: 2.5
    minPrice:
      name: minPrice
      in: query
      required: true
      schema:
        type: number
        example: 0.0
    maxPrice:
      name: maxPrice
      in: query
      required: true
      schema:
        type: number
        example: 5.0
    score:
      name: score
      in: path
      required: true
      schema:
        type: number
        example: 1.5
    minScore:
      name: minScore
      in: query
      required: true
      schema:
        type: number
        example: 1.0
    maxScore:
      name: maxScore
      in: query
      required: true
      schema:
        type: number
        example: 7.0
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT